It starts working again. No conj() or -matlab_val is necessary! This mixes stuff... Just get raw-values from MATLAB.
Make sure that the background-value is subtracted the way, that REAL/IMAG is zero in regions, where there is no scattering! Otherwise the Born won't be able to recover it- or it takes very long (global offset!) Zero-initialization works okish. Still some weird contrast inversion in the Tikhonov reconstruction. Maybe missed some "-" in the derivation? TV shoudl be working. Still need to look at the gradient though!